backend/bedrock-agent/travelAgent/pyproject.toml

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "travelAgent"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "aws-opentelemetry-distro >= 0.10.0",
    "bedrock-agentcore >= 1.0.3",
    "jinja2 >= 3.1.0",
    "pydantic >= 2.0.0",
    "python-dotenv >= 1.2.1",
    "strands-agents[openai] >= 1.13.0",
]